server/requirements.txt
fastapi
uvicorn
# google-genai
requests
python-dotenv
pydantic
ruff
